Clean and re-encode a vector, list or dataframe

A function to re-encode Greek (and other) characters and to remove symbols at the end of a record.

cln(x, level = 1, what, na.rm)

Arguments

x
a vector, list or dataframe

level
optional clean level, either 0 for no-clean, default 1 or 2 (see details)

what
additional characters to clean, optional

na.rm
Remove entries with NA data? (optional and logical)

Details

This function is meant to re-encode Greek (and other) characters in the EDH set given either as list format, vector, or a dataframe produced with function edhw() for example. By default, symbols "?" "*" "+" placed at the end of each record are removed after the re-encoding. However, when level is 0 only re-enmoding is performed, and level 2 is either to force an extra iteration in the re-encoding or when what is invoked.

Value

Depending on the input, a vector, list or dataframe.

See also

Examples

# clean Greek characters
cln("Caesar?*+")

#> [1] "Caesar*+"