Organize EDH dataset province and dates by similarity

Wrapper function to organize EDH dataset province and dates by simple match similarity.

edhwpd(x = NULL, vars, province, dates, clean)

Arguments

x

typically fragments of EDH dataset or database API (optional, list)

vars

variables or attributes to be chosen from x (vector)

province

choosen EDH province abbreviation

dates

vector with TAQ and TPQ (optional)

clean

whether to remove special characters in text (optional and logical)

Details

As with function edhw, this is an interface to extract attribute variables in vars from the EDH or similar dataset if x is not specified. The aim is to organize data per province and dates by simple match similarity among records.

The Roman Empire province is the abbreviation used in the value given by function get.edh and which is in "rp" dataset.

Argument dates is optional to specify the variables for time intervals (TAQ and TPQ) that in EDH are "not_after" and "not_before".

Argument clean applies function cln to the province data frame with the chosen variables to remove special characters such as ?*+ and, if needed, re-encode the text.

The output is a list of data frames with similar arguments by descending matches. The records with one or less similarity matches (or having NA attribute values) are placed in the last data frame of the list.

Value

List of data frames of a EDH class object that includes the province with the number of records.

Author

Antonio Rivero Ostoic

See also

Examples

if (FALSE) {
## load dataset
data(EDH)

## extract province & dates with a single variable attribute from EDH
edhwpd(vars="type_of_inscription", province="Rom", dates=c("not_after", "not_before"))
}