Get data from the Epigraphic Database Heidelberg API

A function to obtain data from the Epigraphic Database Heidelberg new API repository.

get.edh(search = c("inscriptions", "geography"), url =
                 "https://edh.ub.uni-heidelberg.de/data/api", hd_nr,
                 province, country, findspot_modern, findspot_ancient,
                 year_not_before, year_not_after, tm_nr, transcription,
                 type, bbox, findspot, pleiades_id, geonames_id,
                 offset, limit, maxlimit = 4000, addID, printQ)

Arguments

search

Whether the search is on inscriptions or on geography.

url

Open data repository API

hd_nr

HD number of inscription

province

Ancient Roman province name

country

Actual country name

findspot_modern

Actual location name findspot

findspot_ancient

Ancient location name findspot

year_not_before

Year, not before (integer, BC years are negative)

year_not_after

Year, not after (integer, BC years are negative)

tm_nr

Trismegistos ID (integer)

transcription

Automatic leading and trailing truncation (brackets are ignored)

type

Type of inscription (case insensitive)

bbox

Bounding box with character format bbox = "minLong, minLat, maxLong, maxLat"

findspot

Level of village, street etc. (add leading and/or trailing)

pleiades_id

Pleiades identifier of a place (integer)

geonames_id

Geonames identifier of a place (integer)

offset

Clause to specify which row to start from retrieving data (optional and integer)

limit

Clause to limit the number of results (optional and integer)

maxlimit

Maximum limit of the query (integer, default 4000)

addID

Add identification to the output? (optional and logical)

printQ

Print also query? (optional and logical)

Details

Since with the inscriptions option the id "component" of the output list is not with a numeric format, then the function adds an ID at the beginning of the list with the identifier with a numerical format.

Notice that hd_nr is not the same as ID nor id.

Use function get.edhw in case you want to grab several items.

A list with the of valid values from the EDH API for the ancient Roman provinces are

"Ach"	Achaia	"Cor"	Corsica	"Mes"	Mesopotamia

"Aeg"	Aegyptus	"Cre"	Creta	"MoI"	Moesia inferior

"Aem"	Aemilia (Regio VIII)	"Cyp"	Cyprus	"MoS"	Moesia superior

"Afr"	Africa Proconsularis	"Cyr"	Cyrene	"Nar"	Narbonensis

"AlC"	Alpes Cottiae	"Dac"	Dacia	"Nor"	Noricum

"AlG"	Alpes Graiae	"Dal"	Dalmatia	"Num"	Numidia

"AlM"	Alpes Maritimae	"Epi"	Epirus	"PaI"	Pannonia inferior

"AlP"	Alpes Poeninae	"Etr"	Etruria (Regio VII)	"PaS"	Pannonia superior

"ApC"	Apulia et Calabria (Regio II)	"Gal"	Galatia	"Pic"	Picenum (Regio V)

"Aqu"	Aquitania	"GeI"	Germania inferior	"Rae"	Raetia

"Ara"	Arabia	"GeS"	Germania superior	"ReB"	Regnum Bospori

"Arm"	Armenia	"HiC"	Hispania citerior	"Rom"	Roma

"Asi"	Asia	"Inc"	Provincia incerta	"Sam"	Samnium (Regio IV)

"Ass"	Assyria	"Iud"	Iudaea	"Sar"	Sardinia

"Bae"	Baetica	"LaC"	Latium et Campania (Regio I)	"Sic"	Sicilia, Melita

"Bar"	Barbaricum	"Lig"	Liguria (Regio IX)	"Syr"	Syria

"Bel"	Belgica	"Lug"	Lugdunensis	"Thr"	Thracia

"BiP"	Bithynia et Pontus	"Lus"	Lusitania	"Tra"	Transpadana (Regio XI)

"BrL"	Bruttium et Lucania (Regio III)	"LyP"	Lycia et Pamphylia	"Tri"	Tripolitania

"Bri"	Britannia	"MaC"	Mauretania Caesariensis	"Umb"	Umbria (Regio VI)

"Cap"	Cappadocia	"MaT"	Mauretania Tingitana	"Val"	Valeria

"Cil"	Cilicia	"Mak"	Macedonia	"VeH"	Venetia et Histria (Regio X)

And the valid values for country entries are abbreviated country names where the inscription was located.

"ad"	Andorra	"gr"	Greece	"pl"	Poland

"al"	Albania	"hr"	Croatia	"pt"	Portugal

"am"	Armenia	"hu"	Hungary	"rks"	Kosovo

"at"	Austria	"il"	Israel	"ro"	Romania

"az"	Azerbaijan	"iq"	Iraq	"rs"	Serbia

"ba"	Bosnia and Herzegovina	"it"	Italy	"ru"	Russia

"be"	Belgium	"jo"	Jordan	"sa"	Saudi Arabia

"bg"	Bulgaria	"kg"	Kyrgyzstan	"sd"	Sudan

"ch"	Switzerland	"kz"	Kazakhstan	"se"	Sweden

"cy"	Cyprus	"lb"	Lebanon	"si"	Slovenia

"cz"	Czech Republic	"li"	Liechtenstein	"sk"	Slovakia

"de"	Germany	"lu"	Luxembourg	"sm"	San Marino

"dk"	Denmark	"ly"	Libyan Arab Jamahiriya	"sy"	Syrian Arab Republic

"dz"	Algeria	"ma"	Morocco	"tj"	Tajikistan

"eg"	Egypt	"mc"	Monaco	"tn"	Tunisia

"es"	Spain	"md"	Moldova	"tr"	Turkey

"fr"	France	"me"	Montenegro	"ua"	Ukraine

"gb"	United Kingdom	"mk"	Macedonia	"uz"	Uzbekistan

"ge"	Georgia	"mt"	Malta	"va"	Vatican City State

"gi"	Gibraltar	"nl"	Netherlands	"ye"	Yemen

Value

A list object with at least one the following items:

"commentary"

"fotos"

"country"

"depth"

"diplomatic_text"

"edh_geography_uri"

"findspot"

"findspot_ancient"

"findspot_modern"

"geography"

"height"

"id"

"language"

"last_update"

"letter_size"

"literature"

"material"

"military"

"modern_region"

"not_after"

"not_before"

"people"

This item is another list with at least one the following items:

"person_id"
"nomen"
"cognomen"
"praenomen"
"name"
"gender"
"status"
"tribus"
"origo"
"occupation"
"age: years"
"age: months"
"age: days"

"present_location"

"religion"

"province_label"

"responsible_individual"

"social_economic_legal_history"

"transcription"

"trismegistos_uri"

"type_of_inscription"

"type_of_monument"

"uri"

"width"

"work_status"

"year_of_find"

"ID"

(Optional), only if addID is set to TRUE.

And the query is printed if specified by printQ.

References

https://edh.ub.uni-heidelberg.de/data/api

Author

Antonio Rivero Ostoic

Note

This function is for the [EDH] database [API] starting in year January 2022.

Warning

For queries having more than 4000 records, which can produce a timeout from the server.

See also

Examples

## Not run:
# get inscriptions from EDH API data
get.edh(findspot_modern="madrid")
#> [[1]]
#> [[1]]$country
#> [1] "Spain"
#> 
#> [[1]]$diplomatic_text
#> [1] "SINEVRSA / ET QVINTILLVS / FILIVS AN XX / H S SVNT"
#> 
#> [[1]]$findspot_modern
#> [1] "Madrid?"
#> 
#> [[1]]$id
#> [1] "HD015327"
#> 
#> [[1]]$language
#> [1] "L"
#> 
#> [[1]]$last_update
#> [1] "1997-06-03"
#> 
#> [[1]]$letter_size
#> NULL
#> 
#> [[1]]$literature
#> [1] "AE 1989, 0361d. ;"
#> 
#> [[1]]$responsible_individual
#> [1] "Scheithauer"
#> 
#> [[1]]$transcription
#> [1] "Sineursa(?) / et Quintillus / filius an(norum) XX / h(ic) s(iti) sunt"
#> 
#> [[1]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/236310"
#> 
#> [[1]]$work_status
#> [1] "provisional"
#> 
#> 
#> [[2]]
#> [[2]]$country
#> [1] "Spain"
#> 
#> [[2]]$diplomatic_text
#> [1] "L IVNIVS / EPAGATVS / AN XI / H S E S T T L"
#> 
#> [[2]]$findspot_modern
#> [1] "Madrid?"
#> 
#> [[2]]$id
#> [1] "HD015330"
#> 
#> [[2]]$language
#> [1] "L"
#> 
#> [[2]]$last_update
#> [1] "1997-06-03"
#> 
#> [[2]]$letter_size
#> NULL
#> 
#> [[2]]$literature
#> [1] "AE 1989, 0361e. ;"
#> 
#> [[2]]$responsible_individual
#> [1] "Scheithauer"
#> 
#> [[2]]$transcription
#> [1] "L(ucius) Iunius / Epagatus / an(norum) XI / h(ic) s(itus) e(st) s(it) t(ibi) t(erra) l(evis)"
#> 
#> [[2]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/236309"
#> 
#> [[2]]$work_status
#> [1] "provisional"
#> 
#> 
#> [[3]]
#> [[3]]$commentary
#> [1] "Verschollen. Mögliche Datierung: 99-100."
#> 
#> [[3]]$country
#> [1] "Spain"
#> 
#> [[3]]$diplomatic_text
#> [1] "[ ] / [ ] / [ ] / GER PO[ ]TIF / [ ] / [ ] / [ ] / ["
#> 
#> [[3]]$findspot
#> [1] "Puerta de Moros"
#> 
#> [[3]]$findspot_modern
#> [1] "Madrid"
#> 
#> [[3]]$id
#> [1] "HD041220"
#> 
#> [[3]]$language
#> [1] "L"
#> 
#> [[3]]$last_update
#> [1] "2020-09-01"
#> 
#> [[3]]$letter_size
#> NULL
#> 
#> [[3]]$literature
#> [1] "J. Lostal Pros, Los miliarios de la provincia tarraconense (conventos tarraconense, cesaraugustano, cluniense y cartaginense) (Zaragoza 1992) 70, Nr. 65. ; CIL 02, 03056. ;"
#> 
#> [[3]]$responsible_individual
#> [1] "Garcia"
#> 
#> [[3]]$transcription
#> [1] "[Imp(erator) Nerva] / [Caesar Aug(ustus)] / [Traianus] / Ger(manicus) po[n]tif(ex) / [max(imus) trib(unicia)] / [pot(estate) IIII p(ater) p(atriae)] / [co(n)s(ul) II] / [------"
#> 
#> [[3]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/228383"
#> 
#> [[3]]$type_of_inscription
#> [1] "mile-/leaguestone"
#> 
#> [[3]]$type_of_monument
#> [1] "mile-/leaguestone"
#> 
#> [[3]]$work_status
#> [1] "no image"
#> 
#> 
## End(Not run)