Wrapper to get data from the Epigraphic Database Heidelberg API

A wrapper function to obtain data from the Epigraphic Database Heidelberg repository.

get.edhw(file = NULL, hd_nr, ...)

Arguments

file
JSON file with EDH data (optional)

hd_nr
HD number of inscriptions

...
Additional arguments

Details

This is a wrapper function to obtain a sample data from the Epigraphic Database Heidelberg API repository by their HD numbers or, alternatively, a file with a valid format JSON can be specified in file.

In any case, the JSON output will be converted into a list object with the rjson package.

Value

A list of lists object with the items described in get.edh.

References

https://edh-www.adw.uni-heidelberg.de/data/api

Note

Large samples can take a lot of time.

See also

Examples

## get 10 records from EDH API data
## Not run:
get.edhw(hd_nr=1:10)

#> [[1]]
#> [[1]]$ID
#> [1] "000001"
#> 
#> [[1]]$commentary
#> [1] " (C): 2. Hälfte 1. - Anfang 2. Jh. - AE; Ende 1. - 1. Hälfte 2. Jh. - Annecchino."
#> 
#> [[1]]$country
#> [1] "Italy"
#> 
#> [[1]]$depth
#> [1] "2.7 cm"
#> 
#> [[1]]$diplomatic_text
#> [1] "D M / NONIAE P F OPTATAE / ET C IVLIO ARTEMONI / PARENTIBVS / LIBERTIS LIBERTABVSQVE / POSTERISQVE EORVM / C IVLIVS C F OPTATVS / FILIVS"
#> 
#> [[1]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/11843"
#> 
#> [[1]]$findspot_ancient
#> [1] "Cumae, bei"
#> 
#> [[1]]$findspot_modern
#> [1] "Cuma, bei"
#> 
#> [[1]]$height
#> [1] "33 cm"
#> 
#> [[1]]$id
#> [1] "HD000001"
#> 
#> [[1]]$language
#> [1] "Latin"
#> 
#> [[1]]$last_update
#> [1] "2014-04-07"
#> 
#> [[1]]$letter_size
#> [1] "3.2-2 cm"
#> 
#> [[1]]$literature
#> [1] "AE 1983, 0192.; M. Annecchino, Puteoli 4/5, 1980/81, 286-287, Nr. 17; fig. 17. (C) - AE 1983."
#> 
#> [[1]]$material
#> [1] "Marmor, geädert / farbig"
#> 
#> [[1]]$modern_region
#> [1] "Campania"
#> 
#> [[1]]$not_after
#> [1] "0130"
#> 
#> [[1]]$not_before
#> [1] "0071"
#> 
#> [[1]]$people
#> [[1]]$people[[1]]
#> [[1]]$people[[1]]$nomen
#> [1] "Nonia"
#> 
#> [[1]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[1]]$people[[1]]$cognomen
#> [1] "Optata"
#> 
#> [[1]]$people[[1]]$name
#> [1] "Noniae P.f. Optatae"
#> 
#> [[1]]$people[[1]]$gender
#> [1] "female"
#> 
#> 
#> [[1]]$people[[2]]
#> [[1]]$people[[2]]$cognomen
#> [1] "Artemo"
#> 
#> [[1]]$people[[2]]$gender
#> [1] "male"
#> 
#> [[1]]$people[[2]]$praenomen
#> [1] "C."
#> 
#> [[1]]$people[[2]]$name
#> [1] "C. Iulio Artemoni"
#> 
#> [[1]]$people[[2]]$nomen
#> [1] "Iulius"
#> 
#> [[1]]$people[[2]]$person_id
#> [1] "2"
#> 
#> 
#> [[1]]$people[[3]]
#> [[1]]$people[[3]]$gender
#> [1] "male"
#> 
#> [[1]]$people[[3]]$praenomen
#> [1] "C."
#> 
#> [[1]]$people[[3]]$name
#> [1] "C. Iulius C.f. Optatus"
#> 
#> [[1]]$people[[3]]$cognomen
#> [1] "Optatus"
#> 
#> [[1]]$people[[3]]$person_id
#> [1] "3"
#> 
#> [[1]]$people[[3]]$nomen
#> [1] "Iulius"
#> 
#> 
#> 
#> [[1]]$province_label
#> [1] "Latium et Campania (Regio I)"
#> 
#> [[1]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[1]]$transcription
#> [1] "D(is) M(anibus) / Noniae P(ubli) f(iliae) Optatae / et C(aio) Iulio Artemoni / parentibus / libertis libertabusque / posterisque eorum / C(aius) Iulius C(ai) f(ilius) Optatus / filius"
#> 
#> [[1]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/251193"
#> 
#> [[1]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[1]]$type_of_monument
#> [1] "tabula"
#> 
#> [[1]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000001"
#> 
#> [[1]]$width
#> [1] "34 cm"
#> 
#> [[1]]$work_status
#> [1] "provisional"
#> 
#> 
#> [[2]]
#> [[2]]$ID
#> [1] "000002"
#> 
#> [[2]]$commentary
#> [1] " AE 1983: Breite: 35 cm."
#> 
#> [[2]]$country
#> [1] "Italy"
#> 
#> [[2]]$diplomatic_text
#> [1] "C SEXTIVS PARIS / QVI VIXIT / ANNIS LXX"
#> 
#> [[2]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/19482"
#> 
#> [[2]]$findspot
#> [1] "Via Nomentana, S. Alessandro, Kirche"
#> 
#> [[2]]$findspot_ancient
#> [1] "Roma"
#> 
#> [[2]]$findspot_modern
#> [1] "Roma"
#> 
#> [[2]]$height
#> [1] "28 cm"
#> 
#> [[2]]$id
#> [1] "HD000002"
#> 
#> [[2]]$language
#> [1] "Latin"
#> 
#> [[2]]$last_update
#> [1] "2014-04-07"
#> 
#> [[2]]$letter_size
#> [1] "4 cm"
#> 
#> [[2]]$literature
#> [1] "AE 1983, 0080. (A); A. Ferrua, RAL 36, 1981, 109, Nr. 7. - AE 1983."
#> 
#> [[2]]$material
#> [1] "marble: rocks - metamorphic rocks"
#> 
#> [[2]]$modern_region
#> [1] "Lazio"
#> 
#> [[2]]$not_after
#> [1] "0200"
#> 
#> [[2]]$not_before
#> [1] "0051"
#> 
#> [[2]]$people
#> [[2]]$people[[1]]
#> [[2]]$people[[1]]$name
#> [1] "C. Sextius Paris"
#> 
#> [[2]]$people[[1]]$`age: years`
#> [1] "70"
#> 
#> [[2]]$people[[1]]$nomen
#> [1] "Sextius"
#> 
#> [[2]]$people[[1]]$cognomen
#> [1] "Paris"
#> 
#> [[2]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[2]]$people[[1]]$gender
#> [1] "male"
#> 
#> [[2]]$people[[1]]$praenomen
#> [1] "C."
#> 
#> 
#> 
#> [[2]]$province_label
#> [1] "Roma"
#> 
#> [[2]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[2]]$transcription
#> [1] "C(aius) Sextius Paris / qui vixit / annis LXX"
#> 
#> [[2]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/265631"
#> 
#> [[2]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[2]]$type_of_monument
#> [1] "tabula"
#> 
#> [[2]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000002"
#> 
#> [[2]]$width
#> [1] "85 cm"
#> 
#> [[2]]$work_status
#> [1] "no image"
#> 
#> [[2]]$year_of_find
#> [1] "1937"
#> 
#> 
#> [[3]]
#> [[3]]$ID
#> [1] "000003"
#> 
#> [[3]]$commentary
#> [1] " (B): [S]isenna ist falscher Kasus; folgende Ergänzung notwendig: [S]isenna[e]. Y: Text bezieht sich eher auf P. Mummius P.f. Sisenna Rutilianus (cos. suff. 146) als auf dessen Vater P. Mummius Sisenna (cos. ord. 133). Die Statuenbasis war in einem Gebäude vermauert."
#> 
#> [[3]]$country
#> [1] "Spain"
#> 
#> [[3]]$depth
#> [1] "(12) cm"
#> 
#> [[3]]$diplomatic_text
#> [1] "[ ]VMMIO [ ] / [ ]ISENNA[ ] / [ ] XV[ ] / [ ] / ["
#> 
#> [[3]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/2694"
#> 
#> [[3]]$findspot_modern
#> [1] "Tomares"
#> 
#> [[3]]$height
#> [1] "(37) cm"
#> 
#> [[3]]$id
#> [1] "HD000003"
#> 
#> [[3]]$language
#> [1] "Latin"
#> 
#> [[3]]$last_update
#> [1] "2006-08-31"
#> 
#> [[3]]$letter_size
#> [1] "4.5-3 cm"
#> 
#> [[3]]$literature
#> [1] "AE 1983, 0518. (B); J. González, ZPE 52, 1983, 172-173; Taf. 5b. (B) - AE."
#> 
#> [[3]]$material
#> [1] "marble: rocks - metamorphic rocks"
#> 
#> [[3]]$modern_region
#> [1] "Sevilla"
#> 
#> [[3]]$not_after
#> [1] "0170"
#> 
#> [[3]]$not_before
#> [1] "0131"
#> 
#> [[3]]$people
#> [[3]]$people[[1]]
#> [[3]]$people[[1]]$praenomen
#> [1] "P.+"
#> 
#> [[3]]$people[[1]]$gender
#> [1] "male"
#> 
#> [[3]]$people[[1]]$status
#> [1] "senatorial order"
#> 
#> [[3]]$people[[1]]$nomen
#> [1] "Mummius+"
#> 
#> [[3]]$people[[1]]$cognomen
#> [1] "Sisenna+ Rutilianus+"
#> 
#> [[3]]$people[[1]]$tribus
#> [1] "Galeria+"
#> 
#> [[3]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[3]]$people[[1]]$name
#> [1] "[P. M]ummio [P.f. Gal. S]isenna[e Rutiliano]"
#> 
#> 
#> 
#> [[3]]$present_location
#> [1] "Sevilla, Privatbesitz"
#> 
#> [[3]]$province_label
#> [1] "Baetica"
#> 
#> [[3]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[3]]$transcription
#> [1] "[P(ublio) M]ummio [P(ubli) f(ilio)] / [Gal(eria) S]isenna[e Ru]/[tiliano] Xv[ir(o) stli]/[tibus iudicandis ---] / [------"
#> 
#> [[3]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/220675"
#> 
#> [[3]]$type_of_inscription
#> [1] "honorific inscription"
#> 
#> [[3]]$type_of_monument
#> [1] "statue base"
#> 
#> [[3]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000003"
#> 
#> [[3]]$width
#> [1] "(34) cm"
#> 
#> [[3]]$work_status
#> [1] "provisional"
#> 
#> [[3]]$year_of_find
#> [1] "before 1975"
#> 
#> 
#> [[4]]
#> [[4]]$ID
#> [1] "000004"
#> 
#> [[4]]$commentary
#> [1] " Material: lokaler grauer Kalkstein. (B): Stylow, AE 1983: Z. 1: A[---]L(?)L(?)A; CIL II 1638: Z. 1: Aemilia; Z. 2: Mar[--] Iponubensis(?); Z. 3: Veneris."
#> 
#> [[4]]$country
#> [1] "Spain"
#> 
#> [[4]]$depth
#> [1] "18 cm"
#> 
#> [[4]]$diplomatic_text
#> [1] "[ ]AVS[ ]LLA / M PORCI NIGRI SER / DOMINAE VENERI / ARAM POSVIT"
#> 
#> [[4]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/2240"
#> 
#> [[4]]$external_image_uris
#> [1] "http://cil-old.bbaw.de/test06/bilder/datenbank/PH0005311.jpg"
#> 
#> [[4]]$findspot_ancient
#> [1] "Ipolcobulcula"
#> 
#> [[4]]$findspot_modern
#> [1] "Carcabuey"
#> 
#> [[4]]$height
#> [1] "(39) cm"
#> 
#> [[4]]$id
#> [1] "HD000004"
#> 
#> [[4]]$language
#> [1] "Latin"
#> 
#> [[4]]$last_update
#> [1] "2015-03-27"
#> 
#> [[4]]$letter_size
#> [1] "2.5 cm"
#> 
#> [[4]]$literature
#> [1] "AE 1983, 0533. (B); A.U. Stylow, Gerión 1, 1983, 267-273, Nr. 1; fig. 1. - AE 1983.; CIL 02, 01638. (B); CIL 02 (2. Aufl.) 05, 00274; Foto."
#> 
#> [[4]]$material
#> [1] "limestone: rocks - clastic sediments"
#> 
#> [[4]]$modern_region
#> [1] "Córdoba"
#> 
#> [[4]]$not_after
#> [1] "0200"
#> 
#> [[4]]$not_before
#> [1] "0151"
#> 
#> [[4]]$people
#> [[4]]$people[[1]]
#> [[4]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[4]]$people[[1]]$status
#> [1] "slaves"
#> 
#> [[4]]$people[[1]]$cognomen
#> [1] "[---]"
#> 
#> [[4]]$people[[1]]$name
#> [1] "[---?]AV(?)S(?)[---]L(?)L(?)A M. Porci Nigri ser."
#> 
#> 
#> [[4]]$people[[2]]
#> [[4]]$people[[2]]$cognomen
#> [1] "Niger"
#> 
#> [[4]]$people[[2]]$person_id
#> [1] "2"
#> 
#> [[4]]$people[[2]]$praenomen
#> [1] "M."
#> 
#> [[4]]$people[[2]]$name
#> [1] "M. Porci Nigri"
#> 
#> [[4]]$people[[2]]$nomen
#> [1] "Porcius"
#> 
#> 
#> 
#> [[4]]$present_location
#> [1] "Carcabuey, Grupo Escolar"
#> 
#> [[4]]$province_label
#> [1] "Baetica"
#> 
#> [[4]]$religion
#> [1] "names of pagan deities"
#> 
#> [[4]]$responsible_individual
#> [1] "Gräf"
#> 
#> [[4]]$transcription
#> [1] "[---?]AV(?)S(?)[---]L(?)L(?)A / M(arci) Porci Nigri ser(va) / dominae Veneri / aram posuit"
#> 
#> [[4]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/222102"
#> 
#> [[4]]$type_of_inscription
#> [1] "votive inscription"
#> 
#> [[4]]$type_of_monument
#> [1] "altar"
#> 
#> [[4]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000004"
#> 
#> [[4]]$width
#> [1] "27 cm"
#> 
#> [[4]]$work_status
#> [1] "checked with photo"
#> 
#> [[4]]$year_of_find
#> [1] "before 1979"
#> 
#> 
#> [[5]]
#> [[5]]$ID
#> [1] "000005"
#> 
#> [[5]]$commentary
#> [1] " (B): Z. 3: C(ai) l(ibertae) Tyches."
#> 
#> [[5]]$country
#> [1] "Italy"
#> 
#> [[5]]$diplomatic_text
#> [1] "[ ] L SVCCESSVS / [ ] L L IRENAEVS / [ ] C L TYCHES / [ ]VNT / [ ]RENTI F"
#> 
#> [[5]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/19428"
#> 
#> [[5]]$findspot
#> [1] "Via Cupa (ehem. Vigna Nardi)"
#> 
#> [[5]]$findspot_ancient
#> [1] "Roma"
#> 
#> [[5]]$findspot_modern
#> [1] "Roma"
#> 
#> [[5]]$id
#> [1] "HD000005"
#> 
#> [[5]]$language
#> [1] "Latin"
#> 
#> [[5]]$last_update
#> [1] "2010-01-04"
#> 
#> [[5]]$literature
#> [1] "AE 1983, 0078. (B); A. Ferrua, RAL 36, 1981, 108, Nr. 3. - AE 1983."
#> 
#> [[5]]$modern_region
#> [1] "Lazio"
#> 
#> [[5]]$not_after
#> [1] "0200"
#> 
#> [[5]]$not_before
#> [1] "0001"
#> 
#> [[5]]$people
#> [[5]]$people[[1]]
#> [[5]]$people[[1]]$praenomen
#> [1] "[-]"
#> 
#> [[5]]$people[[1]]$gender
#> [1] "male"
#> 
#> [[5]]$people[[1]]$name
#> [1] "[---]l. Successus"
#> 
#> [[5]]$people[[1]]$nomen
#> [1] "[---]"
#> 
#> [[5]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[5]]$people[[1]]$status
#> [1] "freedmen / freedwomen"
#> 
#> [[5]]$people[[1]]$cognomen
#> [1] "Successus"
#> 
#> 
#> [[5]]$people[[2]]
#> [[5]]$people[[2]]$gender
#> [1] "male"
#> 
#> [[5]]$people[[2]]$praenomen
#> [1] "[-]"
#> 
#> [[5]]$people[[2]]$name
#> [1] "[---] L.l. Irenaeus"
#> 
#> [[5]]$people[[2]]$nomen
#> [1] "[---]"
#> 
#> [[5]]$people[[2]]$person_id
#> [1] "2"
#> 
#> [[5]]$people[[2]]$status
#> [1] "freedmen / freedwomen"
#> 
#> [[5]]$people[[2]]$cognomen
#> [1] "Irenaeus"
#> 
#> 
#> [[5]]$people[[3]]
#> [[5]]$people[[3]]$name
#> [1] "[---] C.l. Tyches(!)"
#> 
#> [[5]]$people[[3]]$nomen
#> [1] "[---]"
#> 
#> [[5]]$people[[3]]$person_id
#> [1] "3"
#> 
#> [[5]]$people[[3]]$gender
#> [1] "female"
#> 
#> [[5]]$people[[3]]$cognomen
#> [1] "Tyche"
#> 
#> [[5]]$people[[3]]$status
#> [1] "freedmen / freedwomen"
#> 
#> 
#> 
#> [[5]]$province_label
#> [1] "Roma"
#> 
#> [[5]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[5]]$transcription
#> [1] "[---] l(ibertus) Successus / [---] L(uci) l(ibertus) Irenaeus / [---] C(ai) l(iberta) Tyches(!) / [---]unt / [--- me]renti f(---)"
#> 
#> [[5]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/265629"
#> 
#> [[5]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[5]]$type_of_monument
#> [1] "stele"
#> 
#> [[5]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000005"
#> 
#> [[5]]$work_status
#> [1] "no image"
#> 
#> 
#> [[6]]
#> [[6]]$ID
#> [1] "000006"
#> 
#> [[6]]$commentary
#> [1] " Der Stein ist aus 2 aneinanderpassenden Fragmenten zusammengesetzt. Z. 1 befindet sich oberhalb des gerahmten Inschriftfeldes. (B): Rodríguez Oliva - Atencia, AE 1983: Z. 4: t(erra)."
#> 
#> [[6]]$country
#> [1] "Spain"
#> 
#> [[6]]$depth
#> [1] "15 cm"
#> 
#> [[6]]$diplomatic_text
#> [1] "D M S / / MEMMIA AVCTIN[ ] / AN LXX PIA IN SVIS / H S E S T TE[ ] L"
#> 
#> [[6]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/2999"
#> 
#> [[6]]$findspot
#> [1] "Huerta Nueva"
#> 
#> [[6]]$findspot_ancient
#> [1] "Sabora, bei"
#> 
#> [[6]]$findspot_modern
#> [1] "Cañete la Real"
#> 
#> [[6]]$height
#> [1] "145 cm"
#> 
#> [[6]]$id
#> [1] "HD000006"
#> 
#> [[6]]$language
#> [1] "Latin"
#> 
#> [[6]]$last_update
#> [1] "2011-06-10"
#> 
#> [[6]]$letter_size
#> [1] "7-4 cm"
#> 
#> [[6]]$literature
#> [1] "AE 1983, 0524. (B); P. Rodríguez Oliva - R. Atencia Páez, BSEAA 49, 1983, 154, Nr. 3; lám. 3,2. (B) - AE 1983.; CIL 02 (2. Aufl.) 05, 00883."
#> 
#> [[6]]$material
#> [1] "limestone: rocks - clastic sediments"
#> 
#> [[6]]$modern_region
#> [1] "Málaga"
#> 
#> [[6]]$not_after
#> [1] "0150"
#> 
#> [[6]]$not_before
#> [1] "0071"
#> 
#> [[6]]$people
#> [[6]]$people[[1]]
#> [[6]]$people[[1]]$gender
#> [1] "female"
#> 
#> [[6]]$people[[1]]$cognomen
#> [1] "Auctina+"
#> 
#> [[6]]$people[[1]]$name
#> [1] "Memmia Auctin[a]"
#> 
#> [[6]]$people[[1]]$nomen
#> [1] "Memmia"
#> 
#> [[6]]$people[[1]]$`age: years`
#> [1] "70"
#> 
#> [[6]]$people[[1]]$person_id
#> [1] "1"
#> 
#> 
#> 
#> [[6]]$present_location
#> [1] "Cañete la Real"
#> 
#> [[6]]$province_label
#> [1] "Baetica"
#> 
#> [[6]]$responsible_individual
#> [1] "Gräf"
#> 
#> [[6]]$transcription
#> [1] "D(is) M(anibus) s(acrum) // Memmia Auctin[a] / an(norum) LXX pia in suis / h(ic) s(ita) e(st) s(it) t(ibi) te[r(ra)] l(evis)"
#> 
#> [[6]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/222924"
#> 
#> [[6]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[6]]$type_of_monument
#> [1] "stele"
#> 
#> [[6]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000006"
#> 
#> [[6]]$width
#> [1] "60 cm"
#> 
#> [[6]]$work_status
#> [1] "checked with photo"
#> 
#> [[6]]$year_of_find
#> [1] "1974"
#> 
#> 
#> [[7]]
#> [[7]]$ID
#> [1] "000007"
#> 
#> [[7]]$commentary
#> [1] " (C): Datierung: Zeit Sullas."
#> 
#> [[7]]$country
#> [1] "Italy"
#> 
#> [[7]]$diplomatic_text
#> [1] "CLODIA M F"
#> 
#> [[7]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/19483"
#> 
#> [[7]]$findspot
#> [1] "Via Nomentana, Tor Lupara"
#> 
#> [[7]]$findspot_ancient
#> [1] "Roma"
#> 
#> [[7]]$findspot_modern
#> [1] "Roma"
#> 
#> [[7]]$height
#> [1] "35 cm"
#> 
#> [[7]]$id
#> [1] "HD000007"
#> 
#> [[7]]$language
#> [1] "Latin"
#> 
#> [[7]]$last_update
#> [1] "2010-01-04"
#> 
#> [[7]]$letter_size
#> [1] "9 cm"
#> 
#> [[7]]$literature
#> [1] "AE 1983, 0033. (C); G. Pisani Sartorio, in: G. Barbieri (Hrsg.), Il lapidario Zeri di Mentana (Roma 1982) 3-4, Nr. 2; tav. 2,1. - AE 1983."
#> 
#> [[7]]$material
#> [1] "travertine: rocks - chemische Sedimente"
#> 
#> [[7]]$modern_region
#> [1] "Lazio"
#> 
#> [[7]]$not_after
#> [1] "-0051"
#> 
#> [[7]]$not_before
#> [1] "-0100"
#> 
#> [[7]]$people
#> [[7]]$people[[1]]
#> [[7]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[7]]$people[[1]]$gender
#> [1] "female"
#> 
#> [[7]]$people[[1]]$nomen
#> [1] "Clodia"
#> 
#> [[7]]$people[[1]]$name
#> [1] "Clodia M.f."
#> 
#> 
#> 
#> [[7]]$present_location
#> [1] "Mentana, Privatbesitz"
#> 
#> [[7]]$province_label
#> [1] "Roma"
#> 
#> [[7]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[7]]$transcription
#> [1] "Clodia M(arci) f(ilia)"
#> 
#> [[7]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/265588"
#> 
#> [[7]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[7]]$type_of_monument
#> [1] "tabula"
#> 
#> [[7]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000007"
#> 
#> [[7]]$width
#> [1] "53 cm"
#> 
#> [[7]]$work_status
#> [1] "checked with photo"
#> 
#> 
#> [[8]]
#> [[8]]$ID
#> [1] "000008"
#> 
#> [[8]]$commentary
#> [1] " Tafel aus mehreren anpassenden Fragmenten zusammengesetzt. Vermutlich stadtrömische Inschrift."
#> 
#> [[8]]$country
#> [1] "Italy"
#> 
#> [[8]]$diplomatic_text
#> [1] "D M / C SATRIO XANTHO / C SATRI RVFI LIB / DECVRIALI LICTORIO / SATRIA NYMPHIDIA MARITO / OPTIMO ET FIDELISSIMO ET SIBI ET / LIBERTIS LIBERTABVSQ SVIS / POSTERISQ EORVM FECIT"
#> 
#> [[8]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/19664"
#> 
#> [[8]]$findspot_ancient
#> [1] "Roma?"
#> 
#> [[8]]$findspot_modern
#> [1] "Roma?"
#> 
#> [[8]]$fotos
#> [1] "https://edh-www.adw.uni-heidelberg.de/fotos/F023291.JPG"
#> [2] "https://edh-www.adw.uni-heidelberg.de/fotos/F029577.JPG"
#> 
#> [[8]]$height
#> [1] "52 cm"
#> 
#> [[8]]$id
#> [1] "HD000008"
#> 
#> [[8]]$language
#> [1] "Latin"
#> 
#> [[8]]$last_update
#> [1] "2013-12-03"
#> 
#> [[8]]$letter_size
#> [1] "4-2 cm"
#> 
#> [[8]]$literature
#> [1] "AE 1983, 0060.; R. Palmieri, in: G. Barbieri (Hrsg.), Il lapidario Zeri di Mentana (Roma 1982) 109-110, Nr. 48; tav. 44, 1. - AE 1983."
#> 
#> [[8]]$material
#> [1] "marble: rocks - metamorphic rocks"
#> 
#> [[8]]$modern_region
#> [1] "Lazio?"
#> 
#> [[8]]$not_after
#> [1] "0200"
#> 
#> [[8]]$not_before
#> [1] "0101"
#> 
#> [[8]]$people
#> [[8]]$people[[1]]
#> [[8]]$people[[1]]$status
#> [1] "freedmen / freedwomen"
#> 
#> [[8]]$people[[1]]$name
#> [1] "C. Satrio Xantho C. Satri Rufi lib."
#> 
#> [[8]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[8]]$people[[1]]$gender
#> [1] "male"
#> 
#> [[8]]$people[[1]]$cognomen
#> [1] "Xanthus"
#> 
#> [[8]]$people[[1]]$occupation
#> [1] "data available"
#> 
#> [[8]]$people[[1]]$nomen
#> [1] "Satrius"
#> 
#> [[8]]$people[[1]]$praenomen
#> [1] "C."
#> 
#> 
#> [[8]]$people[[2]]
#> [[8]]$people[[2]]$person_id
#> [1] "2"
#> 
#> [[8]]$people[[2]]$gender
#> [1] "male"
#> 
#> [[8]]$people[[2]]$cognomen
#> [1] "Rufus"
#> 
#> [[8]]$people[[2]]$name
#> [1] "C. Satri Rufi"
#> 
#> [[8]]$people[[2]]$nomen
#> [1] "Satrius"
#> 
#> [[8]]$people[[2]]$praenomen
#> [1] "C."
#> 
#> 
#> [[8]]$people[[3]]
#> [[8]]$people[[3]]$nomen
#> [1] "Satria"
#> 
#> [[8]]$people[[3]]$status
#> [1] "freedmen / freedwomen?"
#> 
#> [[8]]$people[[3]]$name
#> [1] "Satria Nymphidia"
#> 
#> [[8]]$people[[3]]$cognomen
#> [1] "Nymphidia"
#> 
#> [[8]]$people[[3]]$person_id
#> [1] "3"
#> 
#> [[8]]$people[[3]]$gender
#> [1] "female"
#> 
#> 
#> 
#> [[8]]$present_location
#> [1] "Mentana, Privatbesitz"
#> 
#> [[8]]$province_label
#> [1] "Roma?"
#> 
#> [[8]]$responsible_individual
#> [1] "Feraudi"
#> 
#> [[8]]$transcription
#> [1] "D(is) M(anibus) / C(aio) Satrio Xantho / C(ai) Satri Rufi lib(erto) / decuriali lictorio / Satria Nymphidia marito / optimo et fidelissimo et sibi et / libertis libertabusq(ue) suis / posterisq(ue) eorum fecit"
#> 
#> [[8]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/265611"
#> 
#> [[8]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[8]]$type_of_monument
#> [1] "tabula"
#> 
#> [[8]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000008"
#> 
#> [[8]]$width
#> [1] "65 cm"
#> 
#> [[8]]$work_status
#> [1] "provisional"
#> 
#> 
#> [[9]]
#> [[9]]$ID
#> [1] "000009"
#> 
#> [[9]]$commentary
#> [1] " Material: Blei-Zinn-Kupfer-Legierung. Vielleicht zu verstehen als ABC def(i)x(io)."
#> 
#> [[9]]$country
#> [1] "United Kingdom"
#> 
#> [[9]]$diplomatic_text
#> [1] "ABCDEFX"
#> 
#> [[9]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/4502"
#> 
#> [[9]]$findspot
#> [1] "Römisches Bad, Sacred Spring"
#> 
#> [[9]]$findspot_ancient
#> [1] "Aquae Sulis"
#> 
#> [[9]]$findspot_modern
#> [1] "Bath"
#> 
#> [[9]]$height
#> [1] "5.2 cm"
#> 
#> [[9]]$id
#> [1] "HD000009"
#> 
#> [[9]]$language
#> [1] "Latin"
#> 
#> [[9]]$last_update
#> [1] "2019-03-29"
#> 
#> [[9]]$literature
#> [1] "AE 1983, 0632.; M.W.C. Hassall - R.S.O. Tomlin, Britannia 14, 1983, 336, Nr. 2; fig. 32 (Zeichnung). - AE 1983.; R.S.O. Tomlin, in: B. Cunliffe (Hrsg.), The Temple of Sulis Minerva at Bath.; Volume 2 The Finds from the Sacred Spring.; (Oxford 1988) 108, Nr. 1; Zeichnung.; A. Kropp, Defixiones. Ein aktuelles Corpus lateinischer Fluchtafeln (Speyer 2008) Nr. 3.2/3."
#> 
#> [[9]]$material
#> [1] "Blei, Zinn"
#> 
#> [[9]]$modern_region
#> [1] "Somerset"
#> 
#> [[9]]$not_after
#> [1] "0300"
#> 
#> [[9]]$not_before
#> [1] "0201"
#> 
#> [[9]]$present_location
#> [1] "Bath, Roman Baths Mus."
#> 
#> [[9]]$province_label
#> [1] "Britannia"
#> 
#> [[9]]$religion
#> [1] "magic/spells"
#> 
#> [[9]]$responsible_individual
#> [1] "Cowey"
#> 
#> [[9]]$transcription
#> [1] "ABCDEFX"
#> 
#> [[9]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/168722"
#> 
#> [[9]]$type_of_inscription
#> [1] "defixio"
#> 
#> [[9]]$type_of_monument
#> [1] "tessera"
#> 
#> [[9]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000009"
#> 
#> [[9]]$width
#> [1] "6.5 cm"
#> 
#> [[9]]$work_status
#> [1] "checked with drawing"
#> 
#> [[9]]$year_of_find
#> [1] "1979"
#> 
#> 
#> [[10]]
#> [[10]]$ID
#> [1] "000010"
#> 
#> [[10]]$commentary
#> [1] " (B): AE 1983: Z. 3/4: Secundus et Orphaeus."
#> 
#> [[10]]$country
#> [1] "Italy"
#> 
#> [[10]]$depth
#> [1] "44 cm"
#> 
#> [[10]]$diplomatic_text
#> [1] "D M / L ASINI POLI / SECVNDVS / ET ORPHAEVS / LIB P B M"
#> 
#> [[10]]$edh_geography_uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/geographie/259"
#> 
#> [[10]]$findspot
#> [1] "Friedhof"
#> 
#> [[10]]$findspot_ancient
#> [1] "Ariminum"
#> 
#> [[10]]$findspot_modern
#> [1] "Rimini"
#> 
#> [[10]]$height
#> [1] "46 cm"
#> 
#> [[10]]$id
#> [1] "HD000010"
#> 
#> [[10]]$language
#> [1] "Latin"
#> 
#> [[10]]$last_update
#> [1] "2007-05-16"
#> 
#> [[10]]$letter_size
#> [1] "4.2-2.5 cm"
#> 
#> [[10]]$literature
#> [1] "AE 1983, 0410. (B); G.A. Mansuelli, Epigraphica 2, 1940, 185-186, Nr. 2; fig. 9. - AE 1983."
#> 
#> [[10]]$modern_region
#> [1] "Forli"
#> 
#> [[10]]$not_after
#> [1] "0200"
#> 
#> [[10]]$not_before
#> [1] "0101"
#> 
#> [[10]]$people
#> [[10]]$people[[1]]
#> [[10]]$people[[1]]$praenomen
#> [1] "L."
#> 
#> [[10]]$people[[1]]$nomen
#> [1] "Asinius"
#> 
#> [[10]]$people[[1]]$cognomen
#> [1] "Polus"
#> 
#> [[10]]$people[[1]]$name
#> [1] "L. Asini Poli"
#> 
#> [[10]]$people[[1]]$person_id
#> [1] "1"
#> 
#> [[10]]$people[[1]]$gender
#> [1] "male"
#> 
#> 
#> [[10]]$people[[2]]
#> [[10]]$people[[2]]$gender
#> [1] "male"
#> 
#> [[10]]$people[[2]]$status
#> [1] "freedmen / freedwomen"
#> 
#> [[10]]$people[[2]]$name
#> [1] "Secundus"
#> 
#> [[10]]$people[[2]]$person_id
#> [1] "2"
#> 
#> [[10]]$people[[2]]$cognomen
#> [1] "Secundus"
#> 
#> 
#> [[10]]$people[[3]]
#> [[10]]$people[[3]]$status
#> [1] "freedmen / freedwomen"
#> 
#> [[10]]$people[[3]]$gender
#> [1] "male"
#> 
#> [[10]]$people[[3]]$name
#> [1] "Orphaeus"
#> 
#> [[10]]$people[[3]]$cognomen
#> [1] "Orphaeus"
#> 
#> [[10]]$people[[3]]$person_id
#> [1] "3"
#> 
#> 
#> 
#> [[10]]$present_location
#> [1] "Rimini, Mus. Arch. Com."
#> 
#> [[10]]$province_label
#> [1] "Aemilia (Regio VIII)"
#> 
#> [[10]]$responsible_individual
#> [1] "Gräf"
#> 
#> [[10]]$transcription
#> [1] "D(is) M(anibus) / L(uci) Asini Poli / Secundus / et Orphaeus / lib(erti) p(atrono) b(ene) m(erenti)"
#> 
#> [[10]]$trismegistos_uri
#> [1] "https://www.trismegistos.org/text/244297"
#> 
#> [[10]]$type_of_inscription
#> [1] "epitaph"
#> 
#> [[10]]$type_of_monument
#> [1] "urn"
#> 
#> [[10]]$uri
#> [1] "https://edh-www.adw.uni-heidelberg.de/edh/inschrift/HD000010"
#> 
#> [[10]]$width
#> [1] "52.5 cm"
#> 
#> [[10]]$work_status
#> [1] "checked with photo"
#> 
#> [[10]]$year_of_find
#> [1] "1936"
#> 
#> 

## End(**Not run**)