Plot interval dates

A function to plot interval dates.

plot.dates(x, y, type = c("ts", "mp", "rg"), taq, tpq, id, out, 
    col, cex, lwd, lty, pch, main = NULL, xlab = NULL, ylab = NULL, 
    xlim = NULL, axes = TRUE, alpha, file = NULL, ...)

Arguments

x

dataset as a data frame object of variables and observations.

y

optional identifiers (vector).

type

Type of date format to plot:

ts timespans

mp mid points and range

rg only range

taq

terminus ante quem

tpq

terminus post quem

id

IDs as variable or rownames in dataset x

out

number of outliers to omit (integer or vector where first entry id for latest date)

col

color of pch

cex

size of pch

lwd

width of time interval segments

lty

shape of time interval segments

pch

symbol for taq and tpq

main

plot's main tile

xlab

plot's x label

ylab

plot's y label

xlim

plot's x limits

axes

plot's axes (logical)

alpha

alpha transparency for time interval segments

file

path to file for a PDF format (optional)

...

additional optional parameters

Details

This plot function is for time interval segments given in the dataset x, which is given as a data frame or as a ``tibble'' class object.

Value

A graphical plot.

Author

Antonio Rivero Ostoic

Note

If x is NULL, then EDH dataset is taken.

See also

Examples

#TBD.