Digital Tools for the SDAM Project at Aarhus University

Provides digital tools for performing analyses within Social Dynamics and complexity in the Ancient Mediterranean (SDAM), which is a research group based at the Department of History and Classical Studies at Aarhus University.

Details

Package:	sdam

Type:	Package

Version:	1.1.1 (development)

Date:	29 August 2022

License:	CC BY-SA 4.0

The "sdam" package is a suite of tools for performing analyses in the history of the Mediterranean world and its neighbouring regions in the antiquity period before the Middle Ages. Currently, it is possible to access data of writing material from the Epigraphic Database Heidelberg API with function get.edh and the wrapper function get.edhw. Most of the epigraphic data, at least until 10-11-2020, is available in the EDH dataset, which can be transformed in diverse ways by using function edhw and function cln for re-encoding and cleaning portions of text in inscriptions in the EDH dataset.

With function prex, there is also the possibility to compute probabilities of existence of inscriptions and other artefacts with either the aoristic sum or count matching for observations for different periodization options. Function plot.dates allow visualizing interval time events that can be adjusted by the internal function dts as illustrated in a vignette.

request function allows, with user authentication, performing different types of HTTP requests aimed to DEiC's servers in https://sciencedata.dk or another cloud repository with a customized URL address.

A plotting function is found in plot.map that allows visualizing cartographical maps of ancient Roman provinces that are part of the EDH dataset. It relies other datasets from the package that are related to the Roman world in rp, rpd, rpmp, rpmcd, rpcp, and retn for a transport network with terrestrial and maritime main routes.

Similarity by simple matching among column vectors in a table or data frame is achieved by function simil for making analyses of relations between e.g. assemblages and artefacts. Function edhwpd is to organize the EDH dataset per province and dates based on a similarity measure, and it is the basis for function rmids to compute values of missing dates with a restricted imputation on data subsets of artefacts.

Author

Author: Antonio Rivero Ostoic [aut, cre], Adela Sobotkova [ctb], Vojtech Kase [ctb], Petra Hermankova [ctb]

Maintainer: Antonio Rivero Ostoic <jaro@cas.au.dk>

References

Epigraphic Database Heidelberg -- Data Reuse Options, (Online; retrieved on 16 June 2019). URL https://edh-www.adw.uni-heidelberg.de/data

See also

Examples

data(package="sdam")