Wrapper function for the Epigraphic Database Heidelberg dataset

A function to obtain variable data and perform transformations on the EDH dataset.

edhw(vars, x = NULL, as = c("list", "df"), addID, limit, id, na.rm, ...)

Arguments

vars
Choose variables of interest from x. If x=NULL, the entire EDH dataset is taken. (optional, vector)

x
A list object name with fragments of the EDH dataset, or another EDH dataset created either with rjson or jsonlite (optional)

as
Format to return the output. Currently either as a list or a data frame df object.

addID
Add identification to the output? (optional and logical)

limit
Limit the returned output. Ignored if id is specified (optional, integer or vector)

id
Select only the hd_nr id(s) (optional, integer or character)

na.rm
Remove entries with NA data? (logical and optional)

...
Optional arguments if needed.

Details

This is a convenient function to "extract" variables from the EDH dataset attached to this package through vars. When vars is missing, then all entries are taken. Component "people" is a separated list and should be considered as a separate case from the rest of the variables

In case x is present, it is typically given as a fragment of the EDH dataset that can be obtained for example with function get.edh() or get.edhw(). Otherwise, it must be at least a list object with a comparable structure to EDH.

By default, a list object is returned but it is possible to convert the list into a data frame, and with this option the ordering of the variables is given alphabetically. If desired, it is also possible to remove missing data in the output format by activating na.rm.

Arguments id and limit serve to reduce the returned output either to some Epigraphic Database number or numbers, which are specified by hd_nr, or else by limiting the amount of the returned output. This latter is like the limit argument of function get.edh(), where the offset can be specified as well. Argument id is for refering to precisely hd_nrs, while "limit" is a faster way to get to these entries in the EDH dataset.

Value

A list or a data frame, depending on the input.

References

https://edh-www.adw.uni-heidelberg.de/data/api

Note

In case that the output chosen is a dataframe, then component "people" should be chosen alone.

See also

Examples

if (FALSE) {
## load data set
data(EDH)

## make a list for three variables in 'EDH' (default)
edhw(vars=c("ID", "not_after", "not_before") )}