Get data from the Epigraphic Database Heidelberg

A function to obtain data from the Epigraphic Database Heidelberg repository.

get.edh(search = c("inscriptions", "geography"),
        url = "https://edh-www.adw.uni-heidelberg.de/data/api",
        hd_nr, province, country, findspot_modern, findspot_ancient,
        year_not_before, year_not_after, tm_nr, transcription, type,
        bbox, findspot, pleiades_id, geonames_id, offset, limit,
        addID, printQ)

Arguments

search
Whether the search is on inscriptions or on geography.

url
Open data repository API

hd_nr
HD number of inscription

province
Ancient Roman province name

country
Actual country name

findspot_modern
Actual location name findspot

findspot_ancient
Ancient location name findspot

year_not_before
Year, not before (integer, BC years are negative)

year_not_after
Year, not after (integer, BC years are negative)

tm_nr
Trismegistos' database number (?)

transcription
Automatic leading and trailing truncation (brackets are ignored)

type
Type of inscription (case insensitive)

bbox
Bounding box with character format bbox = "minLong, minLat, maxLong, maxLat"

findspot
Level of village, street etc. (add leading and/or trailing)

pleiades_id
Pleiades identifier of a place (integer)

geonames_id
Geonames identifier of a place (integer)

offset
Clause to specify which row to start from retrieving data (optional and integer)

limit
Clause to limit the number of results (optional and integer)

addID
Add identification to the output? (optional and logical)

printQ
Also print query? (optional and logical)

Details

Since with the inscriptions option the id "component" of the output list is not with a numeric format, then the function adds an ID at the beginning of the list with the identifier with a numerical format.

Notice that hd_nr is not the same as ID nor id.

Use function get.edhw in case you want to grab several items.

Entries in country are abbreviated country names where the inscription was located. A list with the of valid values for countries from the EDH API are

"ad"	Andorra	"gr"	Greece	"pl"	Poland

"al"	Albania	"hr"	Croatia	"pt"	Portugal

"am"	Armenia	"hu"	Hungary	"rks"	Kosovo

"at"	Austria	"il"	Israel	"ro"	Romania

"az"	Azerbaijan	"iq"	Iraq	"rs"	Serbia

"ba"	Bosnia and Herzegovina	"it"	Italy	"ru"	Russia

"be"	Belgium	"jo"	Jordan	"sa"	Saudi Arabia

"bg"	Bulgaria	"kg"	Kyrgyzstan	"sd"	Sudan

"ch"	Switzerland	"kz"	Kazakhstan	"se"	Sweden

"cy"	Cyprus	"lb"	Lebanon	"si"	Slovenia

"cz"	Czech Republic	"li"	Liechtenstein	"sk"	Slovakia

"de"	Germany	"lu"	Luxembourg	"sm"	San Marino

"dk"	Denmark	"ly"	Libyan Arab Jamahiriya	"sy"	Syrian Arab Republic

"dz"	Algeria	"ma"	Morocco	"tj"	Tajikistan

"eg"	Egypt	"mc"	Monaco	"tn"	Tunisia

"es"	Spain	"md"	Moldova	"tr"	Turkey

"fr"	France	"me"	Montenegro	"ua"	Ukraine

"gb"	United Kingdom	"mk"	Macedonia	"uz"	Uzbekistan

"ge"	Georgia	"mt"	Malta	"va"	Vatican City State

"gi"	Gibraltar	"nl"	Netherlands	"ye"	Yemen

And for the ancient Roman provices the valued values are

"Ach"	Achaia	"Cor"	Corsica	"Mes"	Mesopotamia

"Aeg"	Aegyptus	"Cre"	Creta	"MoI"	Moesia inferior

"Aem"	Aemilia (Regio VIII)	"Cyp"	Cyprus	"MoS"	Moesia superior

"Afr"	Africa Proconsularis	"Cyr"	Cyrene	"Nar"	Narbonensis

"AlC"	Alpes Cottiae	"Dac"	Dacia	"Nor"	Noricum

"AlG"	Alpes Graiae	"Dal"	Dalmatia	"Num"	Numidia

"AlM"	Alpes Maritimae	"Epi"	Epirus	"PaI"	Pannonia inferior

"AlP"	Alpes Poeninae	"Etr"	Etruria (Regio VII)	"PaS"	Pannonia superior

"ApC"	Apulia et Calabria (Regio II)	"Gal"	Galatia	"Pic"	Picenum (Regio V)

"Aqu"	Aquitania	"GeI"	Germania inferior	"Rae"	Raetia

"Ara"	Arabia	"GeS"	Germania superior	"ReB"	Regnum Bospori

"Arm"	Armenia	"HiC"	Hispania citerior	"Rom"	Roma

"Asi"	Asia	"Inc"	Provincia incerta	"Sam"	Samnium (Regio IV)

"Ass"	Assyria	"Iud"	Iudaea	"Sar"	Sardinia

"Bae"	Baetica	"LaC"	Latium et Campania (Regio I)	"Sic"	Sicilia, Melita

"Bar"	Barbaricum	"Lig"	Liguria (Regio IX)	"Syr"	Syria

"Bel"	Belgica	"Lug"	Lugdunensis	"Thr"	Thracia

"BiP"	Bithynia et Pontus	"Lus"	Lusitania	"Tra"	Transpadana (Regio XI)

"BrL"	Bruttium et Lucania (Regio III)	"LyP"	Lycia et Pamphylia	"Tri"	Tripolitania

"Bri"	Britannia	"MaC"	Mauretania Caesariensis	"Umb"	Umbria (Regio VI)

"Cap"	Cappadocia	"MaT"	Mauretania Tingitana	"Val"	Valeria

"Cil"	Cilicia	"Mak"	Macedonia	"VeH"	Venetia et Histria (Regio X)

Value

A list object with at least one the following items:

"ID"

(Optional), only if addID is set to TRUE.

"commentary"

"fotos"

"country"

"depth"

"diplomatic_text"

"edh_geography_uri"

"findspot"

"findspot_ancient"

"findspot_modern"

"geography"

"height"

"id"

"language"

"last_update"

"letter_size"

"literature"

"material"

"military"

"modern_region"

"not_after"

"not_before"

"people"

This item is another list with at least one the following items:
"person_id"
"nomen"
"cognomen"
"praenomen"
"name"
"gender"
"status"
"tribus"
"origo"
"occupation"
"age: years"
"age: months"

"present_location"

"religion"

"province_label"

"responsible_individual"

"social_economic_legal_history"

"transcription"

"trismegistos_uri"

"type_of_inscription"

"type_of_monument"

"uri"

"width"

"work_status"

"year_of_find"

And also the query is printed if specified.
References

https://edh-www.adw.uni-heidelberg.de/data/api

https://edh-www.adw.uni-heidelberg.de/data/api/terms/country

https://edh-www.adw.uni-heidelberg.de/data/api/terms/province

Note

The other two search options from the [EDH] database [API], which are "photos" and "bibliography" may be implemented in the future.

See also

Examples

# TBD