Plot cartographical map

A function to plot a cartographical map, initially of a Roman province or an Italian region.

plot.map(x = NULL, type = c("plain", "rp", "si", "tetra", "med"), settl, roads, shipr,
         main, cap, date, name, fsize, fcol, fsize2, fcol2, xd, yd, new, ...)

Arguments

x
(character) acronym of ancient Roman province or Italian region (see get.edh)

type
Type of cartographical map:

plain

rp Roman provinces

si Senatorial-Imperial

tetra Tetrarchy

med Mediterranean sea

settl
(logical for cartographical map) Display settlements?

roads
(logical for cartographical map) Display roads?

shipr
(logical for cartographical map) Display shipping routes?

main
(optional for cartographical map) Plot's main title

cap
(logical for provinces) display caption?

date
(logical for provinces) display date?

name
(logical for provinces) display province title name?

fsize
(optional) font size in main title

fcol
(optional) font color in main title

fsize2
(optional) font size in date

fcol2
(optional) font color in date

xd
(optional) x positioning for date

yd
(optional) y positioning for date

new
(optional) whether or not plotted map has superimposed graphics

...
additional optional parameters

Details

This plot function is for creating cartographical maps of ancient provinces and Italian regions of the Roman Empire around year 117 AD. The input data x can be a character vector, but this is intended for a recording output. By default, arguments name and cap are set to TRUE while date is set to FALSE.

The type argument allows plotting cartographical maps related to the Roman Empire and Mediterranean sea.

Value

A cartographical map plot with a name, and a caption with an approximate province establishment date.

Note

Positions for caption and date are fixed for a PDF output.

See also

Examples

#TBD.