Compute probability of existence

A function to compute the probability of existence of events.

prex(x, taq, tpq, vars, bins = NULL, cp, aoristic = TRUE, weight = 1,
     DF, out, plot = FALSE, main = NULL, ...)

Arguments

x
list or data frame object of variables and observations.

taq
terminus ante quem (TAQ)

tpq
terminus post quem (TPQ)

vars
boundaries of existence in x (vector for TAQ and TPQ)

bins
length of the break (integer)

cp
chronological phase

aoristic
return aoristic sum? (logical)

weight
weight to observations

DF
return also data frame with observations? Ignored for plot (logical and optional)

out
number of outliers to omit (integer or vector where first entry id for latest date)

plot
plot the results? (logical and optional)

main
plot's main title (optional)

...
additional optional parameters

Details

Currently, the probability of existence is the aoristic sum computed across events for portions of time delimited by a TAQ in taq and TPQ in tpq, or else by the boundaries of existence in vars.

In case that bins is NULL, then the time breaks take the chronological periods in cp, which by default is "bin5" or five-periods for the EDH dataset. Another built-in option is "bin8" for eight chronological periods, but cp is open for other periodization models as long as they are recorded as a list object.

When aoristic is set to FALSE, then a simple matching of only TAQ and TPQ is computed from x.

Value

A data frame with values according to either bins or cp.

References

For aoristic sum: Crema, E. (2012) ``Modelling temporal uncertainty in archaeological analysis'' J Archaeol Method Theory.

For default chronological periods: see Bevan et al, 2013 (doi: 10.1111/j.1475-4754.2012.00674.x)

See also

Examples

#TBD.