Perform a HTTP request

A function to perform a HTTP request

request(file, URL = "https://sciencedata.dk", method = c("GET", "POST", "PUT", "DELETE"), 
        anonymous = FALSE, cred = NULL, path = "/files", subdomain = NULL, force = FALSE, 
        rm.file, ...)

Arguments

file

the request file

URL

protocol and domain of the url

method

the http verb for the object

anonymous

unauthenticated user? (logical)

cred

authentication credentials (vector with username and password)

path

path to add to the url (optional)

subdomain

subdomain to add to the url (optional)

force

force remote file overwriting? (optional)

rm.file

remove file in local machine? (optional and logical)

...

extra parameters if required

Details

request is basically a HTTP request, first aimed to interact with DEiC's (Danish e-Infrastructure Cooperation) https://sciencedata.dk. However, it is possible to specify the URL path and subdomain if necessary.

There are two types of folders in DEiC's https://sciencedata.dk that are personal and shared folders and both requires authentication with credentials.

The path to the shared folders where the files are located must be specified with the path argument. However, for personal folders is the file argument that includes the path information. Many times, DEiC's https://sciencedata.dk places the data on a subdomain, and for some methods like PUT it is required to specify the subdomain as well.

When a file already exists on the remote server, there is a prompt question for overwriting the file when the PUT method is invoked, and by activating argument force we can prevent confirmation and replace the file.

In case that accessing the server requires basic authentication, then package "tcltk" may be needed as well to input the credentials with a widget prompt, and there is the cred argument for performing a basic authentication without a prompt. Public folders can be accessed as anonymous user.

Value

Depends on the method, an action on the server site. A Response message is returned when the method is PUT with the url and items Date, Status, Content-Type. Currently, method POST is not supported at sciencedata.dk.

Author

Antonio Rivero Ostoic

Note

Aliases for this function are sddk() and SDDK().

See also

Examples

## get a public file from remote server as anonymous user
if (FALSE) {
request("filename.extension", method="GET", anonymous=TRUE)}

## put a file in remote server
if (FALSE) {
request("filename.extension", method="PUT")}

## put an existing file in remote server and force overwriting
if (FALSE) {
request("filename.extension", method="PUT", force=TRUE)}

## put an existing file in remote server and remove file from local machine
if (FALSE) {
request("filename.extension", method="PUT", rm.file=TRUE)}


## remove a file in remote server
if (FALSE) {
request("filename.extension", method="DELETE")}