Digital tools for the SDAM project at Aarhus University

This package provides tools for performing analyses within Social Dynamics and complexity in the Ancient Mediterranean (SDAM), which is a research group based at Aarhus University.

Details

Package:	sdam

Type:	Package

Version:	0.4.0

Date:	22 February 2021

License:	CC BY-SA 4.0

Currently, it is possible with the "sdam" package to access data from the Epigraphic Database Heidelberg API with get.edh(), and the wrapper function get.edhw(). Most of the data is available in the EDH dataset attached to the package, which can be manipulated by using edhw().

Besides, the request() function allows performing different types of HTTP requests to DEiC'S https://sciencedata.dk or another cloud repository with a customized URL address.

There is also the possibility to compute probabilities of existence with prex() with either the aoristic sum or count matching for observations for different periodization options. Function plot.dates() allows visualizing interval time events that can be manipulated by the internal function dts as illustrated in a vignette.

Another plotting function in "sdam" is found in plot.map() that allows visualizing maps of ancient Roman provinces that are found in EDH. This function depends on two datasets that are rpmp .

Similarity by simple matching among column vectors is achieved by the simil() function (still under development) in order to make analyses of relations between assemblages and artifacts.

See also