Similarity between (column) vectors

A function to compute similarity between vectors, which can arise from columns in a data frame or list entries.

simil(x, vars, uniq, diag.incl, dichot, rm.isol, k)

Arguments

x

A list or a data frame object

vars

Column(s) in x representing variable attributes (vector)

uniq

unique entries? (optional and logical)

diag.incl

include entries in matrix diagonal? (optional and logical)

dichot

dichotomize output? (optional and logical)

rm.isol

remove isolates in output? (optional and logical)

k

cut-off for dichotomization. If not specified, max of output.

Details

At this point, the ID column in the input represents the labels of the nodes. In case that an ID column does not exists, then the first column is taken provided that there are not duplicated entry names.

Value

A valued matrix with similarities among units by simple matching.

Author

Antonio Rivero Ostoic

Note

Other similarity measures will be added in the near future.

See also

Examples

# TBD